improvement memory usage
main change: avoid copying individual-specific weights (for logL,grad,H) for each SNP-specific model; instead, keep those weights only in the main data object, and keep track of indices involved in estimation SNP-specific model, to retrieve corresponding weights for that analysis (e.g. when calculating gradient, and so on)

had to add an attribute called mainind to Data class for that purpose, and update methods accordingly. also, to make this work, functions such as CalcLogL() had to be switched back to being methods of the Analyser class